Añadida la limpieza de archivos expirados de la caché en Cron y se modificó el método get() en Cache para aceptar un valor por defecto.

+++ Core/Cache.php
@@ -89,13 +89,18 @@ final class Cache
         }
 
         foreach (scandir($folder) as $fileName) {
+            // saltamos los directorios . y ..
+            if ($fileName === '.' || $fileName === '..') {
+                continue;
+            }
+
             if (filemtime($folder . '/' . $fileName) < time() - self::EXPIRATION) {
                 unlink($folder . '/' . $fileName);
             }
         }
     }
 
-    public static function get(string $key)
+    public static function get(string $key, $default = null)
     {
         // buscamos el archivo y comprobamos su fecha de modificación
         $fileName = self::filename($key);
@@ -105,11 +110,11 @@ final class Cache
             try {
                 return unserialize($data);
             } catch (Throwable $e) {
-                return null;
+                return $default;
             }
         }
 
-        return null;
+        return $default;
     }
 
     public static function has(string $key): bool
